% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqtt-package.R
\docType{package}
\name{mqtt}
\alias{mqtt}
\alias{mqtt-package}
\title{Interoperate with 'MQTT' Message Brokers}
\description{
MQTT' is a machine-to-machine ('M2M')/"Internet of Things" connectivity protocol. It
was designed as an extremely lightweight publish/subscribe messaging transport. It is
useful for connections with remote locations where a small code footprint is required
and/or network bandwidth is at a premium. For example, it has been used in sensors
communicating to a broker via satellite link, over occasional dial-up connections with
healthcare providers, and in a range of home automation and small device scenarios. It
is also ideal for mobile applications because of its small size, low power usage,
minimised data packets, and efficient distribution of information to one or many
receivers. Tools are provided to interoperate with 'MQTT' message brokers in R.
}
\author{
Bob Rudis (bob@rud.is)
}
