% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{mqtt_begin}
\alias{mqtt_begin}
\title{Initiate an MQTT connection}
\usage{
mqtt_begin(mobj)
}
\arguments{
\item{mobj}{an mqtt object created with `mqtt_broker()` or augmented by other functions}
}
\value{
`mobj`
}
\description{
This will initiate a connection to an MQTT broker. You are responsible for saving
the returned object and calling `mqtt_end()` on it to cleanly free up resources.
You are also responsible for managing your own event loop (using `mqtt_loop()`).
}
