% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{mqtt_broker}
\alias{mqtt_broker}
\title{Provide broker connection setup information}
\usage{
mqtt_broker(client_id, host = "test.mosquitto.org", port = 1883L)
}
\arguments{
\item{client_id}{your client id (must be unique)}

\item{host}{broker host/IP}

\item{port}{broker port}
}
\value{
mqtt objec
}
\description{
Provide broker connection setup information
}
