% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribe.R
\name{mqtt_default_connection_callback}
\alias{mqtt_default_connection_callback}
\title{mqtt default connection callback function}
\usage{
mqtt_default_connection_callback(result)
}
\arguments{
\item{result}{the return code of the connection response, one of:
- `0`: success
- `1` : connection refused (unacceptable protocol version)
- `2` : connection refused (identifier rejected)
- `3` : connection refused (broker unavailable)
- `4-255` : reserved for future use}
}
\description{
This will be set by default if no parameter is specified to
connection functions. If you use your own function, it should
be modeled after this one (i.e. take the same function signature).
}
