% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{mqtt_default_disconnection_callback}
\alias{mqtt_default_disconnection_callback}
\title{mqtt default disconnection callback function (does nothing)}
\usage{
mqtt_default_disconnection_callback(result)
}
\arguments{
\item{result}{the return code of the disconnection response (ignored). When
writing your own callback handler, the integer value indicates the reason for
the disconnect.  A value of 0 means the client has requested the disconnection.
Any other value indicates that the disconnect is unexpected.}
}
\description{
mqtt default disconnection callback function (does nothing)
}
