% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{mqtt_default_message_callback}
\alias{mqtt_default_message_callback}
\title{mqtt default message callback function}
\usage{
mqtt_default_message_callback(id, topic, payload, qos, retain)
}
\arguments{
\item{id}{the message id}

\item{topic}{the message topic}

\item{payload}{the message payload (raw)}

\item{qos}{the effective qos for the message}

\item{retain}{is this message marked as "retain"?}
}
\description{
This will be set by default if no parameter is specified to
connection functions. If you use your own function, it should
be modeled after this one (i.e. take the same function signature).\cr
\cr
This default function will display messages and quit after 50 messages have
been received.\cr
\cr
You **must** return a character value (an empty string --- "" --- is fine) from
your callback function.
}
\section{Special Functionality}{

If a message callback function returns the string "`quit`", then the MQTT
connection will be closed. This makes it possible to subscribe to a topic,
keep track of messages and stop processing when a known message is received
or after a certain number of messages (you need to keep track of the latter
yourself).
}

