% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{mqtt_end}
\alias{mqtt_end}
\title{Close an MQTT connection}
\usage{
mqtt_end(mobj)
}
\arguments{
\item{mobj}{an mqtt object created with `mqtt_broker()` or augmented by other functions}
}
\value{
`mobj`
}
\description{
You need to use this if you used `mqtt_begin()`
}
