% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{mqtt_loop}
\alias{mqtt_loop}
\title{Run an mqtt loop iteration}
\usage{
mqtt_loop(mobj, timeout = 1000, max_packets = 1)
}
\arguments{
\item{mobj}{an mqtt object created with `mqtt_broker()` or augmented by other functions}

\item{timeout}{Maximum number of milliseconds to wait for network activity before timing out.
Set to 0 for instant return.  Set negative to use the default of `1000` ms.}

\item{max_packets}{is here for potential forward compatibility but internally it will
be reset to 1.}
}
\value{
`mobj`
}
\description{
Run an mqtt loop iteration
}
