% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{mqtt_run}
\alias{mqtt_run}
\title{Run an MQTT event loop}
\usage{
mqtt_run(mobj, times = 10000, timeout = 1000, max_packets = 1)
}
\arguments{
\item{mobj}{an mqtt object created with `mqtt_broker()` or augmented by other functions}

\item{times}{if `Inf` then run forever, otherwise a positive value which will be executed
this many time. This value is trackes for **all** MQTT exchanges, not just messages.}

\item{timeout}{Maximum number of milliseconds to wait for network activity before timing out.
Set to 0 for instant return.  Set negative to use the default of `1000` ms.}

\item{max_packets}{is here for potential forward compatibility but internally it will
be reset to 1.}
}
\value{
`mobj`
}
\description{
This will provide an event loop that will run `time` number of times or forever (until
a signal is received or the program is terminated).
}
