% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{mqtt_silent_callback}
\alias{mqtt_silent_callback}
\title{mqtt silent callback function (does nothing)}
\usage{
mqtt_silent_callback(...)
}
\arguments{
\item{...}{(ignored)}
}
\description{
mqtt silent callback function (does nothing)
}
