% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribe.R
\name{topic_subscribe}
\alias{topic_subscribe}
\title{Subscribe to a MQTT Topic}
\usage{
topic_subscribe(host = "test.mosquitto.org", port = 1883L,
  client_id = "r_mqtt", topic = "#", keepalive = 60L, qos = 0L,
  message_callback = mqtt_default_message_callback,
  connection_callback = mqtt_default_connection_callback)
}
\arguments{
\item{host, port}{host (character) and port (integer) to connect to. Defaults to
"\code{test.mosquitto.org}".}

\item{client_id}{the client id to use. Defaults to "\code{r_mqtt}".}

\item{topic}{topic to subscribe to. Defaults to wildcard "\code{#}"}

\item{keepalive}{the number of seconds after which the broker should send a PING
message to the client if no other messages have been exchanged in that time.
Defaults to \code{60}}

\item{qos}{integer value \code{0}, \code{1} or \code{2} indicating the Quality of Service to be used for
the message.}
}
\description{
Subscribe to a MQTT Topic
}
