% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{mqtt_silence}
\alias{mqtt_silence}
\title{Silence log and/or error or more callbacks}
\usage{
mqtt_silence(mobj, callbacks = c("error", "log", "publish"))
}
\arguments{
\item{mobj}{an mqtt object created with `mqtt_broker()` or augmented by other functions}

\item{callbacks}{any/all of "`error`","`log`", or "`publish`".}
}
\description{
Silence log and/or error or more callbacks
}
