% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{mqtt_username_pw}
\alias{mqtt_username_pw}
\title{Set username & passwords for the connection}
\usage{
mqtt_username_pw(mobj, username = Sys.get("MQTT_USERNAME"),
  password = Sys.get("MQTT_PASSWORD"))
}
\arguments{
\item{username, password}{auto pulled from the environment (`MQTT_USERNAME`, `MQTT_PASSWORD`)
or specify manually.}
}
\value{
`mobj`
}
\description{
Set username & passwords for the connection
}
